refactor: useMutateReward 분리
