+++ frontend/src/pages/Admin/RewardPage/components/RewardItemList/index.tsx
@@ -1,39 +1,17 @@
-import { useMutation } from '@tanstack/react-query';
-import { useNavigate } from 'react-router-dom';
 import { Reward } from '../../../../../types/domain/reward';
 import { RewardName, RewardItemListContainer, RewardItem } from './style';
-import {
-  CustomerIdParams,
-  MutateReq,
-  RewardIdParams,
-  RewardReqBody,
-} from '../../../../../types/api/request';
-import { patchReward } from '../../../../../api/patch';
-import ROUTER_PATH from '../../../../../constants/routerPath';
 import Button from '../../../../../components/Button';
 import useReward from '../../hooks/useReward';
+import useMutateReward from '../../hooks/useMutateReward';
 
 interface RewardItemListProps {
   cafeId: number;
   customerId: number;
 }
 const RewardItemList = ({ cafeId, customerId }: RewardItemListProps) => {
-  const navigate = useNavigate();
-
   // TODO: 내카페의 고객이 아닌 고객의 리워드를 조회할 경우 ErrorMessage를 띄어줌
   const { data: rewardData, status: rewardStatus } = useReward(cafeId, customerId);
-
-  const { mutate: mutateReward } = useMutation({
-    mutationFn: (request: MutateReq<RewardReqBody, RewardIdParams & CustomerIdParams>) => {
-      return patchReward(request);
-    },
-    onSuccess() {
-      navigate(ROUTER_PATH.customerList);
-    },
-    onError() {
-      alert('에러가 발생했습니다. 네트워크 상태를 확인해주세요.');
-    },
-  });
+  const { mutate: mutateReward } = useMutateReward();
 
   const activateRewardButton = (rewardId: number) => {
     mutateReward({
